+++ Turbo_Wheels/Controllers/UsersController.cs
@@ -1,26 +1,23 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Data;
-using System.Data.Entity;
-using System.Linq;
+﻿using System.Linq;
 using System.Net;
-using System.Web;
+using System.Web.Helpers;
 using System.Web.Mvc;
+using Turbo_Wheels.Filters;
 using Turbo_Wheels.Models;
 
 namespace Turbo_Wheels.Controllers
 {
+    [RequireLogin]
+    [RequireAdmin]
     public class UsersController : Controller
     {
         private Turbo_WheelsContext db = new Turbo_WheelsContext();
 
-        // GET: Users
         public ActionResult Index()
         {
             return View(db.Users.ToList());
         }
 
-        // GET: Users/Details/5
         public ActionResult Details(int? id)
         {
             if (id == null)
@@ -35,47 +32,38 @@ namespace Turbo_Wheels.Controllers
             return View(user);
         }
 
-        // GET: Users/Create
         public ActionResult Create()
         {
-            if (Request.Cookies["isAdmin"] == null)
-            {
-                Response.Cookies["isAdmin"].Value = false.ToString();
-            }
             return View();
         }
 
-        // POST: Users/Create
-        // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
-        // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Create([Bind(Include = "userID,username,password,isAdmin,firstName,lastName,email,address,phone")] User user)
+        public ActionResult Create(User user)
         {
-            if (ModelState.IsValid)
+            if (!ModelState.IsValid)
+                return View(user);
+
+            bool userExists = db.Users.Any(u => u.Username == user.Username);
+
+            if (userExists)
             {
-                User user1 = (from item in db.Users where item.username == user.username select item).FirstOrDefault();
-                if (user1==null)
-                {
-                    if (bool.Parse(Request.Cookies["isAdmin"].Value) == true)
-                    {
-                        db.Users.Add(user);
-                        db.SaveChanges();
-                        return RedirectToAction("Index");
-                    }
-                    else
-                    {
-                        user.isAdmin = false;
-                        db.Users.Add(user);
-                        db.SaveChanges();
-                        return RedirectToAction("Login", "Home");
-                    }
-                }   
+                ModelState.AddModelError("Username", "Username already exists.");
+                return View(user);
             }
-            return View(user);
+
+            // Trim and normalize username and email
+            user.Username = user.Username.Trim();
+            user.Email = user.Email.Trim().ToLowerInvariant();
+
+            // Hash the password before saving
+            user.Password = Crypto.HashPassword(user.Password);
+
+            db.Users.Add(user);
+            db.SaveChanges();
+            return RedirectToAction("Index");
         }
 
-        // GET: Users/Edit/5
         public ActionResult Edit(int? id)
         {
             if (id == null)
@@ -90,43 +78,64 @@ namespace Turbo_Wheels.Controllers
             return View(user);
         }
 
-        // POST: Users/Edit/5
-        // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
-        // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Edit([Bind(Include = "userID,username,password,isAdmin,firstName,lastName,email,address,phone")] User user)
+        public ActionResult Edit(User user)
         {
-            if (ModelState.IsValid)
+            if (!ModelState.IsValid)
+                return View(user);
+
+            var existingUser = db.Users.Find(user.UserID);
+            if (existingUser == null)
+                return HttpNotFound();
+
+            // Update fields (explicitly)
+            existingUser.Username = user.Username;
+            existingUser.IsAdmin = user.IsAdmin;
+            existingUser.FirstName = user.FirstName;
+            existingUser.LastName = user.LastName;
+            existingUser.Email = user.Email.Trim().ToLowerInvariant();
+            existingUser.Address = user.Address;
+            existingUser.Phone = user.Phone;
+
+            // Only update password if a new one was provided
+            if (!string.IsNullOrWhiteSpace(user.Password))
             {
-                db.Entry(user).State = EntityState.Modified;
-                db.SaveChanges();
-                return RedirectToAction("Index");
+                existingUser.Password = Crypto.HashPassword(user.Password);
             }
-            return View(user);
+
+            db.SaveChanges();
+            return RedirectToAction("Index");
         }
 
-        // GET: Users/Delete/5
         public ActionResult Delete(int? id)
         {
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
+
             User user = db.Users.Find(id);
+
             if (user == null)
             {
                 return HttpNotFound();
             }
+
             return View(user);
         }
 
-        // POST: Users/Delete/5
-        [HttpPost, ActionName("Delete")]
+        [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult DeleteConfirmed(int id)
+        public ActionResult Delete(int id)
         {
             User user = db.Users.Find(id);
+
+            if (user == null)
+            {
+                return HttpNotFound();
+            }
+
             db.Users.Remove(user);
             db.SaveChanges();
             return RedirectToAction("Index");
